+++ Scripts/Core/Coroutine/KiCoroutine.cs
@@ -9,15 +9,16 @@ namespace KimicuUtility
 
         /// <summary> Start new coroutine. </summary>
         /// <param name="forced"> Forcibly stopping the previous routine if it is already running. </param>
-        public void StartRoutine(IEnumerator enumerator, bool forced = false)
+        public bool StartRoutine(IEnumerator enumerator, bool forced = false)
         {
             if (_routine != null)
             {
                 if (forced) StopRoutine();
-                else return;
+                else return false;
             }
 
             _routine = KiCoroutines.StartRoutine(enumerator);
+            return true;
         }
 
         /// <summary> Stop current coroutine. </summary>
@@ -25,6 +26,7 @@ namespace KimicuUtility
         {
             if (_routine == null) return;
             KiCoroutines.StopRoutine(_routine);
+            _routine = null;
         }
     }
 }