+++ CHANGELOG.md
@@ -5,3 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.0.17] - 06.10.2023
+### Changed
+- Now the "StartRoutine" method in KiCoroutine returns bool.
+
+### Fixed
+- Fix RecurringDelay. The delay did not work correctly.